refactor(embedded-servers): extract embedded servers domain into a zustand slice
Move the embedded HTTP/FTP/TFTP server domain (state + CRUD/lifecycle
actions) out of the monolithic appStore.ts into a dedicated StateCreator
slice at src/store/slices/embedded-serversSlice.ts, combined into the root
store via spread. Mirrors the SSH tunnel slice extraction (#2077, PR #2112).

Moved verbatim: embeddedServers, embeddedServerStates and
loadEmbeddedServers, refreshEmbeddedServerStates, saveEmbeddedServer,
deleteEmbeddedServer, startEmbeddedServer, stopEmbeddedServer,
updateEmbeddedServerState, quickShareServer. AppState now extends
EmbeddedServersSlice, so the public store shape and every action signature
and behavior are unchanged.

Pure structural move — zero behavior change, no consumer changes. The
tab-opening actions and their createTab factory stay in the root store per
the openTunnelEditorTab note in PR #2112.
